feat: rename ArenaLeaderboardApit to GraphLeaderboard
