+++ src/Frontend/Views/Portal/MapDetails.tsx
@@ -1,6 +1,6 @@
 import { GraphConfigPlayer, Leaderboard, LiveMatch } from '@darkforest_eth/types';
 import React, { useEffect, useState } from 'react';
-import { loadArenaLeaderboard } from '../../../Backend/Network/GraphApi/ArenaLeaderboardApi';
+import { loadArenaLeaderboard } from '../../../Backend/Network/GraphApi/GrandPrixLeaderboard';
 import {
   loadEloLeaderboard,
 } from '../../../Backend/Network/GraphApi/EloLeaderboardApi';
